\name{ACGaussian}
\alias{ACGaussian}

\title{Albert and Chib algorithm with Gaussian distribution for latent variable}

\description{Implementation of Albert and Chib (1993) algorithm as described by Sorensen et. al (1995) 
				for using in categorical data analysis}
				
\details{Internal use}

\references{

 ALBERT, J.H.; CHIB, S. (1993) Bayesian analysis of binary anmd polychotomus response data. Journal 
	American Statistical Association, Washington, v.88, n.422, p.669-679.

	SORENSEN, D.A.; ANDERSEN S.; GIANOLA D.; KORSGAARD I.(1995) Bayesian inference in threshold models using
	Gibbs sampling, Genetics Selection Evolution, Les Ulis, v. 27, n. 3, p. 229-249.
}
\seealso{\code{\link{Bayesthresh}}}
