\name{ACt}
\alias{ACt}

\title{Albert and Chib algorithm with t-Student distribution for latent variable}

\description{Algorithm described by Silva and Bueno Filho (2010) for using in categical data analysis adapted
	from Albert and Chib (1993).}

\details{Internal use}

\references{

	ALBERT, J.H.; CHIB, S. (1993) Bayesian analysis of binary and polychotomus response data. Journal 
	American Statistical Association, Washington, v.88, n.422, p.669-679.

	SILVA, W.J,; BUENO FILHO. (2010) An algorithm for thresholds models using cumulative Gaussian or
	cummulative Student's "t" distributions. Brazilian Journal of Biometry, Sao Paulo, v. 28, n. 3, p.59-83. 
}
\seealso{\code{\link{Bayesthresh}}}
\keyword{models}
