\name{MCGaussian}
\alias{MCGaussian}

\title{Kizilkaya et. al. (2003) algorithm with Gaussian distribution for latent variable}

\description{Implementation follows Kizilkaya et. al. (2003), that presented an algorithm
				adapted from Cowles (1996) for mixed models with Gaussian distribution of the latent variable.
  }

\details{Internal use}

\references{
				COWLES, M. K.  (1996) Accelerating Monte Carlo Markov chain convergence for cumulative link	
				generalized linear models, Statistics and Computing,Netherlands: Springer, v.6, n. 2 p. 101-111.
				
				KIZILKAYA, K.; CARNIER, P.; ALBERA, A.; BITTANTE, G.; TEMPELMAN,R. J. (2003) Cumulative t-link 
				threshold models for genetic analysis of calving ease scores, Genetics Selection Evolution,	
				Les Ulis, v. 35, p. 489-512.
}
\seealso{\code{\link{Bayesthresh}}}
