\name{MCt}
\alias{MCt}

\title{Kizilkaya et. al. (2003) algorithm using Student's "t" distribution as link function}

\description{
				The implementation follows Kizilkaya et. al. (2003) which presented a modification on Cowles (1996)
				algorithm that allows for using a Student's "t" distribution as link function for the latent variable. 
}

\details{Internal use}

\references{
				COWLES, M. K. Accelerating Monte Carlo Markov chain convergence for cumulative link 
				generalized linear models, Statistics and Computing,Netherlands: Springer, v.6, n. 2
				p. 101-111, June 1996.
				
				KIZILKAYA, K.; CARNIER, P.; ALBERA, A.; BITTANTE, G.; TEMPELMAN,R. J. Cumulative t-link 
				threshold models for genetic analysis of calving ease scores, Genetics Selection Evolution,
				Les Ulis, v. 35, p. 489-512, Mar. 2003.
}
\seealso{\code{\link{Bayesthresh}}}
