\name{NCGaussian}
\alias{NCGaussian}

\title{Nandram and Chen (1996) algorithm with Gaussian distribution for latent variable}

\description{
				This brings an implementation of Nandram and Chen (1996) algorithm adapted by Silva (2008)
				for mixed models with inverse Gaussian distribution used as a link function for the latent variable}
				
\details{Internal use}

\references{
				NANDRAM, B.; CHEN, M. H. (1996) Reparameterizing the generalized linear model to
				accelerate gibbs sampler convergence, Journal of Statistical Computation and 
				Simulation, Abingdon, v. 54, p.129-144.
				
				SILVA, J.W. (2008) Algorithms for thresholds models using Normal and Student-t cumulative distributions.
				Thesis (Doctorate in Statistics and Agricultural Experimentation), University Federal of Lavras, p. 99.
}
\seealso{\code{\link{Bayesthresh}}}
