\name{NCt}
\alias{NCt}

\title{Nandram and Chen (1996) algorithm with t-Student distribution for latent variable}
	
\description{
				Implementation of Nandram and Chen (1996) algorithm using inverse Student's "t" distribution
				as link function for latent variable (adapted by Silva 2008).
			}

\details{Internal use}

\references{
				NANDRAM, B.; CHEN, M. H. (1996) Reparameterizing the generalized linear model to
				accelerate gibbs sampler convergence, Journal of Statistical Computation and
				Simulation, Abingdon, v. 54, p.129-144.
			
				SILVA, J.W. (2008) Algorithms for thresholds models using Normal and Student-t cumulative distributions.
				Thesis (Doctorate in Statistics and Agricultural Experimentation), University Federal of Lavras, p. 99.
}
\seealso{\code{\link{Bayesthresh}}}
