\name{sensory}
\alias{sensory}

\title{Sensory analysis of banana}

\description{An data set for illustrated of the use package \code{Bayesthresh}}

\usage{data(sensory)
       sensory
}

\details{
				Data used by Silva and Bueno-Filho (2010) for illustrated the using of the algorithm ACG and ACt.
}

\references{
	SILVA, W.J,; BUENO FILHO. (2010) An algorithm for thresholds models using cumulative Gaussian or
	cummulative Student's "t" distributions. Brazilian Journal of Biometry, Sao Paulo, v. 28, n. 3, p.59-83. 

}


\keyword{models}
\keyword{methods}
