\name{summary.Bayesthresh}
\alias{summary.Bayesthresh}

\title{Summary}
\description{
  Summary of Bayesthresh
}
\usage{
	\method{summary}{Bayesthresh}(object, ...)
}
\arguments{
  \item{object}{
    object class of \code{"Bayesthresh"}
   }
  \item{...}{
   no usage
  }
}
\keyword{models}
\keyword{methods}
